modified:   docker-compose.yml
Updated the Docker Compose file to allow for
more containers in testing, including single database
and multiple databases. Failure to get a valid password
(RANDOM ROOT PASSWORD) and the use of either MARIADB
or MYSQL variables where appropriate

modified:   db-auto-backup.py
Updated the Python script to allow for graceful Failure
when there is no valid password for the database rather
than crashing the script and not backing up anything
on any valid container.

MariaDB and MySQL can now handle the case of there being
a blank root password

MariaDB and MySQL can now handle the case of there being
a specific database specified in the environment variables
not requiring the root password to be know, in this case
only the backup of the database specfied in the environment
variables will be backed up

modified:   requirements.txt
Locked requests to 2.31.0 to avoid issues with newer versions
specifically it breaks the docker socket connection
docker library and this library will both have to be
upgrade simultaneously to avoid issues